pyproject.toml
[project]
name = "wallet"
version = "0.1.0"
description = "Projeto para cumprir o desafio para a vaga de desenvolvedor backend sênior da IplanRio."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.29",
  "sqlalchemy>=2.0",
  "asyncpg>=0.29",
  "psycopg2-binary>=2.9.9",
  "alembic>=1.13",
  "pydantic-settings>=2.2",
  "python-multipart>=0.0.9",
  "passlib[bcrypt]>=1.7",
  "bcrypt==4.0.1",
  "python-jose>=3.3.0",
  "pyjwt>=2.8",
  "authlib[firebase]>=1.3",
  "pyotp>=2.9",
  "slowapi[redis]>=0.1",
  "redis>=5.0",
  "prometheus-fastapi-instrumentator>=6.1.0",
  "prometheus-client>=0.19.0",
  "aiofiles>=23.2.1",
  "numpy>=1.26.4",
  "httpx>=0.27.0",
  "pydantic[email]>=2.11.4",
  "pytest>=8.3.5",
  "pytest-asyncio>=0.23",
  "httpx[async]>=0.27",
  "fakeredis[aioredis]>=2.21",
  "aiosqlite>=0.21.0",
  "google-genai>=1.11.0",
]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
